Revert "Move WorkflowTaskExecException and WorkflowDataException into the top level exceptions (and out of BPMN package).  These have no direct dependency on BPMN and can, at times, produce better and more informed error messages."
This reverts commit 4efd016501ee18f2504d761f33e2b8058e1fb97e.
