+++ SpiffWorkflow/bpmn/specs/ScriptTask.py
@@ -21,7 +21,7 @@ import logging
 from .BpmnSpecMixin import BpmnSpecMixin
 from ...task import TaskState
 from ...specs.Simple import Simple
-from SpiffWorkflow.exceptions import WorkflowTaskExecException
+from SpiffWorkflow.bpmn.exceptions import WorkflowTaskExecException
 
 
 class ScriptTask(Simple, BpmnSpecMixin):
