+++ SpiffWorkflow/bpmn/specs/ServiceTask.py
@@ -5,7 +5,7 @@ import logging
 from .BpmnSpecMixin import BpmnSpecMixin
 from ...task import TaskState
 from ...specs.Simple import Simple
-from SpiffWorkflow.exceptions import WorkflowTaskExecException
+from SpiffWorkflow.bpmn.exceptions import WorkflowTaskExecException
 
 LOG = logging.getLogger(__name__)
