+++ SpiffWorkflow/specs/LoopResetTask.py
@@ -17,6 +17,7 @@
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA
 # 02110-1301  USA
 
+
 from .base import TaskSpec
 from ..task import TaskState
 from SpiffWorkflow.exceptions import WorkflowException
