+++ tests/SpiffWorkflow/camunda/InvalidBusinessRuleTaskParserTest.py
@@ -1,7 +1,7 @@
 import os
 import unittest
 
-from SpiffWorkflow.exceptions import WorkflowTaskExecException
+from SpiffWorkflow.bpmn.exceptions import WorkflowTaskExecException
 from SpiffWorkflow import TaskState
 from SpiffWorkflow.bpmn.workflow import BpmnWorkflow
